(feat) migration script refactor
