+++ hummingbot/client/config/conf_migration.py
@@ -11,10 +11,12 @@ import yaml
 from hummingbot import root_path
 from hummingbot.client.config.config_crypt import BaseSecretsManager, store_password_verification
 from hummingbot.client.config.config_data_types import BaseConnectorConfigMap
-from hummingbot.client.config.config_helpers import ClientConfigAdapter
+from hummingbot.client.config.config_helpers import ClientConfigAdapter, save_to_yml
 from hummingbot.client.config.security import Security
-from hummingbot.client.config.xemm_migration import migrate_xemm_confs
 from hummingbot.client.settings import CONF_DIR_PATH, STRATEGIES_CONF_DIR_PATH
+from hummingbot.strategy.cross_exchange_market_making.cross_exchange_market_making_config_map_pydantic import (
+    CrossExchangeMarketMakingConfigMap,
+)
 
 encrypted_conf_prefix = "encrypted_"
 encrypted_conf_postfix = ".json"
@@ -29,7 +31,6 @@ def migrate_configs(secrets_manager: BaseSecretsManager) -> List[str]:
         migrate_strategy_confs_paths()
         errors.extend(migrate_connector_confs(secrets_manager))
         store_password_verification(secrets_manager)
-        migrate_xemm_confs()
         logging.getLogger().info("\nConf migration done.")
     else:
         logging.getLogger().error("\nConf migration failed.")
@@ -75,7 +76,54 @@ def migrate_strategy_confs_paths():
             if "strategy" in conf and _has_connector_field(conf):
                 new_path = strategies_conf_dir_path / child.name
                 child.rename(new_path)
+                if conf["strategy"] == "cross_exchange_market_making":
+                    migrate_xemm_confs()
                 logging.getLogger().info(f"Migrated conf for {conf['strategy']}")
+
+
+def migrate_xemm_confs():
+    logging.getLogger().info("\nMigrating strategies...")
+    for child in strategies_conf_dir_path.iterdir():
+        if child.is_file() and child.name.endswith(".yml"):
+            with open(str(child), "r") as f:
+                conf = yaml.safe_load(f)
+            if "strategy" in conf:
+                if conf["strategy"] == "cross_exchange_market_making":
+                    if "active_order_canceling" in conf:
+                        if conf["active_order_canceling"]:
+                            conf["order_refresh_mode"] = {}
+                        else:
+                            conf["order_refresh_mode"] = {
+                                "cancel_order_threshold": conf["cancel_order_threshold"],
+                                "limit_order_min_expiration": conf["limit_order_min_expiration"]
+                            }
+                        conf.pop("active_order_canceling")
+                        conf.pop("cancel_order_threshold")
+                        conf.pop("limit_order_min_expiration")
+
+                    if "use_oracle_conversion_rate" in conf:
+                        if conf["use_oracle_conversion_rate"]:
+                            conf["conversion_rate_mode"] = {}
+                        else:
+                            conf["conversion_rate_mode"] = {
+                                "taker_to_maker_base_conversion_rate": conf["taker_to_maker_base_conversion_rate"],
+                                "taker_to_maker_quote_conversion_rate": conf["taker_to_maker_quote_conversion_rate"]
+                            }
+                        conf.pop("use_oracle_conversion_rate")
+                        conf.pop("taker_to_maker_base_conversion_rate")
+                        conf.pop("taker_to_maker_quote_conversion_rate")
+
+                    if "template_version" in conf:
+                        conf.pop("template_version")
+
+                    try:
+                        config_map = ClientConfigAdapter(CrossExchangeMarketMakingConfigMap(**conf))
+
+                        save_to_yml(child.absolute(), config_map)
+
+                        logging.getLogger().info(f"Migrated conf for {conf['strategy']}")
+                    except Exception as e:
+                        logging.getLogger().error(str(e))
 
 
 def _has_connector_field(conf: Dict) -> bool:
